Fixed issue with maximised windows.
If maximised windows were snapped there was a chance they would "forget" where
the top of the screen was e.g. when maximised, subsequent to being snapped to
the bottom of the scrren, the window would grow to the correct size to fill the
screen but would continue to place itself halfway downt he screen. Restoring
the window prior to snapping fixes this issue.